requirements.txt
dj_database_url==0.5.0
pyexcel==0.5.14
environs==4.1.0
furl==2.0.0
seqeval==0.0.6
drf_yasg==1.17.1
django_filter==2.0.0
djangorestframework_csv==2.1.0
colour==0.1.5
conllu==1.3.2
apache_libcloud==3.2.0
djangorestframework==3.12.2
